\name{DelayedArray_utils.Rd}

\alias{type,HDF5RealizationSink-method}
\alias{type,arrayRealizationSink-method}
\alias{type,RleRealizationSink-method}

\alias{dimnames,arrayRealizationSink-method}

\title{
Stubs for internal functions
}

\description{Stubs for internal functions}

%\usage{
%dimnames(x)
%type(x)
%}
%\arguments{
%  \item{x}{A RealizationSink (e.g.,
%  arrayRealizationSink, HDF5RealizationSink, RleRealizationSink) instance}
%}

\keyword{internal}
