\name{fixMethOutliers}
\alias{fixMethOutliers}
\title{
  Fix methylation outliers
}
\description{
  Methylation outliers (loci with very extreme values of the Meth or
  Unmeth channel) are identified and fixed (see details).
}
\usage{
fixMethOutliers(object, K = -3, verbose = FALSE)
}
\arguments{
  \item{object}{An object of class \code{[Genomic]MethylSet}.}
  \item{K}{}
  \item{verbose}{Should the function be verbose?}
}
\details{
  
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
}
\author{
  Rafael A. Irizarry
}
\note{
%%  ~~further notes~~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
