\name{getQC}
\alias{getQC}
\alias{addQC}
\alias{plotQC}
\title{
  Estimate sample-specific quality control (QC) for methylation data 
}
\description{
  Estimate sample-specific quality control (QC) for methylation data.
}
\usage{
getQC(object)
addQC(object, qc)
plotQC(qc, badSampleCutoff = 10.5)
}
\arguments{
  \item{object}{An object of class \code{[Genomic]MethylSet}.}
  \item{qc}{}
  \item{badSampleCutoff}{}
}
\details{
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\author{
}
\note{
}
\seealso{
}
\examples{
}
