\name{minfi-defunct}
\alias{minfi-defunct}
\alias{read.450k}
\alias{read.450k.sheet}
\alias{read.450k.exp}
\title{Defunct functions in package \sQuote{minfi}}

\description{
  These functions are provided now defunct in \sQuote{minfi}.
}

\details{
  The following functions are now defunct (not working anymore); use the
  replacement indicated below: 
  \itemize{
    \item{\code{read.450k}: Use \code{\link{read.metharray}}}
    \item{\code{read.450k.sheet}: Use \code{\link{read.metharray.sheet}}}
    \item{\code{read.450k.exp}: Use \code{\link{read.metharray.exp}}}
  }
}

\seealso{
  \code{\link{Defunct}}.
}