\name{minfi-deprecated}
\alias{minfi-deprecated}
\title{Deprecated functions in package \sQuote{minfi}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{minfi} only, and will be defunct at the next release.
}

\details{
  No functions are currently deprecated.
  
  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{
    \item{\code{read.450k}: \code{\link{read.metharray}}}
    \item{\code{read.450k.sheet}: \code{\link{read.metharray.sheet}}}
    \item{\code{read.450k.exp}: \code{\link{read.metharray.exp}}}
  }
}