\name{minfi-package}
\alias{minfi-package}
\alias{minfi}
\docType{package}
\title{
  Analyze Illumina's methylation arrays
}
\description{
  Tools for analyzing and visualizing Illumina methylation array data.
  There is special focus on the 450k array; the 27k array is not
  supported at the moment.
}
\details{
  The package contains a (hopefully) useful vignette; this vignette
  contains a lengthy description of the package content and
  capabilities.
}
\keyword{ package }
