\name{logit2}
\alias{logit2}
\alias{ilogit2}
\title{
  logit in base 2.
}
\description{
  Utility functions for computing logit and inverse logit in base 2.
}
\usage{
logit2(x)
ilogit2(x)
}
\arguments{
  \item{x}{A numeric vector.}
}
\value{A numeric vector.}
\author{
  Kasper Daniel Hansen \email{khansen@jhsph.edu}.
}
\examples{
logit2(c(0.25, 0.5, 0.75))
}
