\name{minfiQC}
\alias{minfiQC}
\title{
  FIXME
}
\description{
}
\usage{
minfiQC(object, fixOutliers = TRUE, verbose = FALSE)
}
\arguments{
  \item{object}{An object of class \code{[Genomic]MethylSet}.}
  \item{fixOutliers}{Should the function fix outlying observations
    (using \code{fixMethOutliers}) before running QC?}
  \item{verbose}{Should the function be verbose?}
}
\details{
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\author{
}
\note{
}
\seealso{
  \code{\link{getSex}}, \code{\link{getQC}}, \code{\link{fixMethOutliers}}
}
\examples{
}
